Add .keep file to empty directory
When trying to deploy the follow error happened to me:

```
Errno::ENOENT: No such file or directory @ dir_initialize - /tmp/build_00ed1b07/app/assets/javascripts/libraries
```

It was because the folder was empty thus not committed with Git.
This commit just adds the file, so we always ensure the folder is committed.