Standard Library of ARIA controllers
Inspired by [Experiment with a standard library of
controllers][stdlib-pr], introduce some Stimulus controllers that
implement keyboard interactions for ARIA patterns in accordance with
their specifications.

While the stimulus-rails repository might not be the ultimate
destination for code like this, being a Rails engine makes it possible
for the JavaScript controllers to be covered by Selenium-powered
Sysytem Tests, and provides an opportunity to be published alongside
corresponding Rails View Helpers, Capybara selectors, and System Test
assertions and helpers.

Ultimately, maybe the controller source code can be published alongside
`hotwired/stimulus`, and the View Helpers, Selectors, and System Test
methods can remain in this gem.

[stdlib-pr]: https://github.com/hotwired/stimulus-rails/pull/24

aria--expander
---

Opens `<dialog>` and `<details>` elements, toggles the
`aria--expander-toggle-class` when provided, and falls back to toggling
the `[hidden]` attribute for all other circumstances.

aria--dialog
---

While the `<dialog>` element serves as a foundation, there are
[additional considerations to the ARIA specifications for
role="dialog"][role="dialog"]. For example, role="dialog" elements are
required to needs to trap focus navigation to interactive elements
within:

> Like non-modal dialogs, modal dialogs contain their tab sequence. That
> is, <kbd>Tab</kbd> and <kbd>Shift + Tab</kbd> do not move focus
> outside the dialog. However, unlike most non-modal dialogs, modal
> dialogs do not provide means for moving keyboard focus outside the
> dialog window without closing the dialog.
>
> When a dialog opens, focus moves to an element inside the dialog. See
> notes below regarding initial focus placement.
>
> It is strongly recommended that the tab sequence of all dialogs
> include a visible element with role button that closes the dialog,
> such as a close icon or cancel button.

The additional behavior covering this is implemented within the
`aria--dialog` controller, which supports the `<dialog>` element.

[dialog]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/dialog
[role="dialog"]: https://www.w3.org/TR/wai-aria-practices/#dialog_modal

aria--feed
---

Navigates focus across [role="article"] or `<article>` descendants of a
container element annotated with [role="feed"][]:

> A feed is a section of a page that automatically loads new sections of
> content as the user scrolls. The sections of content in a feed are
> presented in article elements. So, a feed can be thought of as a dynamic
> list of articles that often appears to scroll infinitely.

[role="feed"]: https://www.w3.org/TR/wai-aria-practices/#feed