Make autoloading changes visible to the DOM
Closes [#35][]
Related to [#49][]

Extend the autoloader to communicate autoloading state via
a [DOMTokenList][] stored in the `[data-stimulus-autoloading]`
attribute.

The goal is to communicate state to the DOM so that tools like Capybara
can monitor the presence or absence of the attribute, and synchronize
page interactions based on whether autoloading is in-flight. By waiting,
our System Tests can act with confidence that any lazily-loaded behavior
is ready.

[#35]: https://github.com/hotwired/stimulus-rails/issues/35
[#49]: https://github.com/hotwired/stimulus-rails/pull/49/files#r610601520
[DOMTokenList]: https://developer.mozilla.org/en-US/docs/Web/API/DOMTokenList