Update mimemagic
Addresses failure in CI:

```
Fetching gem metadata from https://rubygems.org/............ Your bundle
is locked to mimemagic (0.3.5), but that version could not be found in
any of the sources listed in your Gemfile. If you haven't changed
sources, that means the author of mimemagic (0.3.5) has removed it.
You'll need to update your bundle to a version other than mimemagic
(0.3.5) that hasn't been removed in order to install.
```