🍡 Consitently kebab-case controller names in autoloaders
While transitioning a project from `webpacker` to `importmap-rails` I
discovered a handful of my controllers weren't connecting. The root of
the problem is that filenames are being transformed into registered
controller names differently across `stimulus-autoloader` and
`stimulus-importmap-autoloader`. This impacts controllers that have
three or more words in their name.

```
# For example...
/app/assets/javascripts/controllers/many_much_words_controller.js

# Was being autloaded in the importmap solution as
many-much_words

# While the webpacker solution was
many-much-words
```

This change makes it so that `stimulus-autoloader` and
`stimulus-importmap-autoloader` consistently transform the filename,
easing the transition for any other applications that might have a
`really_long_and_descriptive_controller.js`.