Fixes support for generators as well as manifest and install tasks when working in a Rails engine.
app:stimulus:* tasks correctly run in test/dummy and put files in the right places.
stimulus:* tasks exist and puts files under the engine namespace.
The stimulus controller generator, when run in the engine root, includes the engine namespace.
Namespaces are correctly handled in index.js and in terms of where files are placed.