Rethrow error when failing to register a controller
Fixes https://github.com/hotwired/stimulus-rails/issues/102

When failing to register a controller, Stimulus catches the error and logs a console message. But that means the error won't be logged in a monitoring tool, making it hard to identify the issue.

This change follows the pattern in other files to catch this Type Error, format an error message, and re-throw the error so it can be tracked by monitoring tools.

This is my first contribution to Stimulus, so please let me know what I am missing. Thanks for reviewing it!